+++ _import/tycho-execution/foundry/src/executors/EkuboV3Executor.sol
@@ -150,14 +150,12 @@ contract EkuboV3Executor is IExecutor, ICallback {
         returns (
             TransferManager.TransferType transferType,
             address receiver,
-            address tokenIn,
-            uint256 amount
+            address tokenIn
         )
     {
         // data[36:] skips the 4-byte selector and 32-byte locker id
         bytes calldata payData = data[36:];
 
-        amount = uint128(bytes16(payData[0:16]));
         tokenIn = address(bytes20(payData[36:56]));
         receiver = CORE_ADDRESS;