+++ _import/tycho-execution/foundry/test/TychoRouterSingleSwap.t.sol
@@ -45,6 +45,27 @@ contract FalseTokenCallbackPool is TychoRouterTestSetup {
     }
 }
 
+contract ZeroCallbackAmountPool is TychoRouterTestSetup {
+    // When swap() is called, calls back into TychoRouter with crafted data instructing
+    // the router to transfer zero tokens, attempting to impact delta accounting
+    // in unexpected ways. Pool transfers nothing in return.
+
+    function swap(address, bool, int256, uint160, bytes calldata)
+        external
+        returns (int256, int256)
+    {
+        bytes memory callbackData = abi.encodeWithSignature(
+            "slipstreamsCallback(int256,int256,bytes)",
+            int256(0), // callback amount
+            int256(0),
+            abi.encodePacked(WETH_ADDR, bytes23(0)) // token to steal
+        );
+        (bool success,) = address(msg.sender).call(callbackData);
+        require(success, "Callback failed");
+        return (0, 0);
+    }
+}
+
 contract TychoRouterSingleSwapTest is TychoRouterTestSetup {
     function testSingleSwapPermit2() public {
         // Trade 1 WETH for DAI with 1 swap on Uniswap V2 using Permit2
@@ -188,6 +209,58 @@ contract TychoRouterSingleSwapTest is TychoRouterTestSetup {
         // No USDC was stolen
         assertEq(IERC20(USDC_ADDR).balanceOf(tychoRouterAddr), stolenAmount);
         assertEq(IERC20(USDC_ADDR).balanceOf(address(maliciousPool)), 0);
+    }
+
+    function testZeroCallbackAmountSingleSwap() public {
+        // Alice tries to take advantage of cyclical swap accounting using a malicious
+        // zero-amount callback pool to move router funds into her vault.
+
+        uint256 stolenAmount = 10_000 ether;
+        uint256 routerBalance = 10_000 ether;
+        deal(WETH_ADDR, tychoRouterAddr, stolenAmount);
+
+        ZeroCallbackAmountPool maliciousPool = new ZeroCallbackAmountPool();
+
+        // Malicious pool that gives no tokens and asked for 0 amount in callback
+        deal(WETH_ADDR, address(maliciousPool), stolenAmount);
+
+        // Router has some WETH that does not belong to Alice.
+        deal(WETH_ADDR, tychoRouterAddr, routerBalance);
+
+        bytes memory protocolData = abi.encodePacked(
+            WETH_ADDR, WETH_ADDR, bytes3(0), address(maliciousPool), uint8(1)
+        );
+        bytes memory swap =
+            encodeSingleSwap(address(slipstreamsExecutor), protocolData);
+
+        uint256 aliceVaultBalanceBefore =
+            tychoRouter.balanceOf(ALICE, uint256(uint160(WETH_ADDR)));
+        uint256 routerBalanceBefore =
+            IERC20(WETH_ADDR).balanceOf(tychoRouterAddr);
+
+        vm.startPrank(ALICE);
+
+        vm.expectRevert(
+            abi.encodeWithSelector(TychoRouter__NegativeSlippage.selector, 0, 1)
+        );
+        tychoRouter.singleSwapUsingVault(
+            10_000 ether,
+            WETH_ADDR,
+            WETH_ADDR,
+            1,
+            tychoRouterAddr,
+            noClientFee(),
+            swap
+        );
+
+        uint256 aliceVaultBalanceAfter =
+            tychoRouter.balanceOf(ALICE, uint256(uint160(WETH_ADDR)));
+        uint256 routerBalanceAfter =
+            IERC20(WETH_ADDR).balanceOf(tychoRouterAddr);
+
+        // Alice's should not have changed.
+        assertEq(aliceVaultBalanceBefore, aliceVaultBalanceAfter);
+        assertEq(routerBalanceBefore, routerBalanceAfter);
     }
 
     function testSingleSwapInsufficientApproval() public {