+++ _import/tycho-execution/CHANGELOG.md
@@ -1,3 +1,26 @@
+## [0.172.1](https://github.com/propeller-heads/tycho-execution/compare/0.172.0...0.172.1) (2026-04-07)
+
+
+### Bug Fixes
+
+* don't allow protocols to encode transfer amount ([95fa37c](https://github.com/propeller-heads/tycho-execution/commit/95fa37cba24e022c6a73beadbe9db91e75d238d0))
+* remove zero amount transfer check ([f897776](https://github.com/propeller-heads/tycho-execution/commit/f897776c746f1f7938002d6f4bb058a9c05d8b06))
+* Univ3 cyclic exploit (temp) ([ca96e19](https://github.com/propeller-heads/tycho-execution/commit/ca96e19aa01c6cd5901e0cf6d287250436897ef5))
+
+## [0.172.0](https://github.com/propeller-heads/tycho-execution/compare/0.171.4...0.172.0) (2026-04-07)
+
+
+### Features
+
+* Deploy WETH executor ([#111](https://github.com/propeller-heads/tycho-execution/issues/111)) ([1f64fff](https://github.com/propeller-heads/tycho-execution/commit/1f64fffb0d71830737d4e172ce2fd5ac8d4e434c))
+
+## [0.171.4](https://github.com/propeller-heads/tycho-execution/compare/0.171.3...0.171.4) (2026-04-07)
+
+
+### Bug Fixes
+
+* Zero callback amount exploit ([797f58f](https://github.com/propeller-heads/tycho-execution/commit/797f58f06a06b6669bd3f89ef480373411cf234e))
+
 ## [0.171.3](https://github.com/propeller-heads/tycho-execution/compare/0.171.2...0.171.3) (2026-04-06)